pyproject.toml
[project]
name = "dl-assignment1"
version = "1.0.0"
description = "Deep Learning Assignment 1 — Classification Model Comparison"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
    "transformers>=4.30",
    "datasets>=2.14",
    "open-clip-torch>=2.20",
    "gradio>=4.0",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "tensorboard>=2.14",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "numpy>=1.24",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "black"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
